Moved AnalysisModel and supporting classes into a legacy package.
