+++ org.osate.aadl2.errormodel.analysis/src/org/osate/aadl2/errormodel/analysis/cma/CMAUtils.java
@@ -19,9 +19,9 @@ import org.osate.xtext.aadl2.errormodel.errorModel.ErrorPropagation;
 import org.osate.xtext.aadl2.errormodel.errorModel.SConditionElement;
 import org.osate.xtext.aadl2.errormodel.errorModel.SubcomponentElement;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeSet;
-import org.osate.xtext.aadl2.errormodel.util.AnalysisModel;
+import org.osate.xtext.aadl2.errormodel.legacy.AnalysisModel;
+import org.osate.xtext.aadl2.errormodel.legacy.PropagationPathEnd;
 import org.osate.xtext.aadl2.errormodel.util.EMV2Util;
-import org.osate.xtext.aadl2.errormodel.util.PropagationPathEnd;
 import org.osate.xtext.aadl2.properties.util.GetProperties;
 
 public class CMAUtils {
