+++ org.osate.aadl2.errormodel.analysis/src/org/osate/aadl2/errormodel/analysis/fta/FTAUtils.java
@@ -26,11 +26,11 @@ import org.osate.xtext.aadl2.errormodel.errorModel.SConditionElement;
 import org.osate.xtext.aadl2.errormodel.errorModel.SubcomponentElement;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeSet;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeToken;
-import org.osate.xtext.aadl2.errormodel.util.AnalysisModel;
+import org.osate.xtext.aadl2.errormodel.legacy.AnalysisModel;
+import org.osate.xtext.aadl2.errormodel.legacy.PropagationPathEnd;
 import org.osate.xtext.aadl2.errormodel.util.EM2TypeSetUtil;
 import org.osate.xtext.aadl2.errormodel.util.EMV2Properties;
 import org.osate.xtext.aadl2.errormodel.util.EMV2Util;
-import org.osate.xtext.aadl2.errormodel.util.PropagationPathEnd;
 
 public class FTAUtils {
 	private static AnalysisModel currentAnalysisModel;
