+++ org.osate.aadl2.errormodel.analysis/src/org/osate2/aadl2/errormodel/analysis/prism/Module.java
@@ -27,8 +27,8 @@ import org.osate.xtext.aadl2.errormodel.errorModel.SubcomponentElement;
 import org.osate.xtext.aadl2.errormodel.errorModel.TransitionBranch;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeSet;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeToken;
+import org.osate.xtext.aadl2.errormodel.legacy.PropagationPathEnd;
 import org.osate.xtext.aadl2.errormodel.util.EMV2Util;
-import org.osate.xtext.aadl2.errormodel.util.PropagationPathEnd;
 import org.osate2.aadl2.errormodel.analysis.prism.expression.And;
 import org.osate2.aadl2.errormodel.analysis.prism.expression.Equal;
 import org.osate2.aadl2.errormodel.analysis.prism.expression.Expression;
