+++ org.osate.aadl2.errormodel.analysis/src/org/osate2/aadl2/errormodel/analysis/prism/Util.java
@@ -24,9 +24,9 @@ import org.osate.xtext.aadl2.errormodel.errorModel.EventOrPropagation;
 import org.osate.xtext.aadl2.errormodel.errorModel.RecoverEvent;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeSet;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeToken;
+import org.osate.xtext.aadl2.errormodel.legacy.PropagationPathEnd;
 import org.osate.xtext.aadl2.errormodel.util.EMV2Properties;
 import org.osate.xtext.aadl2.errormodel.util.EMV2Util;
-import org.osate.xtext.aadl2.errormodel.util.PropagationPathEnd;
 import org.osate2.aadl2.errormodel.analysis.prism.expression.Equal;
 import org.osate2.aadl2.errormodel.analysis.prism.expression.Expression;
 import org.osate2.aadl2.errormodel.analysis.prism.expression.Terminal;
