+++ org.osate.xtext.aadl2.errormodel/META-INF/MANIFEST.MF
@@ -31,6 +31,7 @@ Export-Package: org.osate.xtext.aadl2.errormodel,
  org.osate.xtext.aadl2.errormodel.errorModel.impl,
  org.osate.xtext.aadl2.errormodel.errorModel.util,
  org.osate.xtext.aadl2.errormodel.formatting,
+ org.osate.xtext.aadl2.errormodel.legacy,
  org.osate.xtext.aadl2.errormodel.naming,
  org.osate.xtext.aadl2.errormodel.parser.antlr,
  org.osate.xtext.aadl2.errormodel.parser.antlr.internal,
