+++ org.osate.xtext.aadl2.errormodel/src/org/osate/xtext/aadl2/errormodel/legacy/AnalysisModel.java
@@ -1,4 +1,4 @@
-package org.osate.xtext.aadl2.errormodel.util;
+package org.osate.xtext.aadl2.errormodel.legacy;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -27,6 +27,7 @@ import org.osate.xtext.aadl2.errormodel.errorModel.ErrorPropagation;
 import org.osate.xtext.aadl2.errormodel.errorModel.PropagationPath;
 import org.osate.xtext.aadl2.errormodel.errorModel.SubcomponentElement;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeSet;
+import org.osate.xtext.aadl2.errormodel.util.EMV2Util;
 import org.osate.xtext.aadl2.properties.util.InstanceModelUtil;
 
 /**
