+++ org.osate.xtext.aadl2.errormodel/src/org/osate/xtext/aadl2/errormodel/legacy/PropagationGraphBackwardTraversal.java
@@ -1,4 +1,4 @@
-package org.osate.xtext.aadl2.errormodel.util;
+package org.osate.xtext.aadl2.errormodel.legacy;
 
 import java.util.Collection;
 import java.util.LinkedList;
@@ -35,6 +35,9 @@ import org.osate.xtext.aadl2.errormodel.errorModel.QualifiedErrorBehaviorState;
 import org.osate.xtext.aadl2.errormodel.errorModel.SConditionElement;
 import org.osate.xtext.aadl2.errormodel.errorModel.TransitionBranch;
 import org.osate.xtext.aadl2.errormodel.errorModel.TypeSet;
+import org.osate.xtext.aadl2.errormodel.util.EM2TypeSetUtil;
+import org.osate.xtext.aadl2.errormodel.util.EMV2Properties;
+import org.osate.xtext.aadl2.errormodel.util.EMV2Util;
 
 public class PropagationGraphBackwardTraversal {
